+++ src/jvmMain/kotlin/kotools/types/collections/NotEmptyCollection.kt
@@ -8,8 +8,8 @@ import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import kotools.types.core.SinceKotoolsTypes
-import kotools.types.number.PositiveIntJvm
-import kotools.types.number.StrictlyPositiveIntJvm
+import kotools.types.int.PositiveInt
+import kotools.types.int.StrictlyPositiveInt
 import kotools.types.string.NotBlankStringJvm
 
 /**
@@ -30,8 +30,8 @@ public inline fun <E> NotEmptyCollection<E>.getOrElse(
  */
 @SinceKotoolsTypes("1.3")
 public inline fun <E> NotEmptyCollection<E>.getOrElse(
-    index: PositiveIntJvm,
-    defaultValue: (PositiveIntJvm) -> @UnsafeVariance E
+    index: PositiveInt,
+    defaultValue: (PositiveInt) -> @UnsafeVariance E
 ): E = getOrNull(index) ?: defaultValue(index)
 
 /**
@@ -41,8 +41,8 @@ public inline fun <E> NotEmptyCollection<E>.getOrElse(
  */
 @SinceKotoolsTypes("1.3")
 public inline fun <E> NotEmptyCollection<E>.getOrElse(
-    index: StrictlyPositiveIntJvm,
-    defaultValue: (StrictlyPositiveIntJvm) -> @UnsafeVariance E
+    index: StrictlyPositiveInt,
+    defaultValue: (StrictlyPositiveInt) -> @UnsafeVariance E
 ): E = getOrNull(index) ?: defaultValue(index)
 
 /**
@@ -61,8 +61,8 @@ public sealed interface NotEmptyCollection<out E> : Collection<E> {
     override val size: Int get() = 1
 
     /** Returns the [size] of this collection as a strictly positive int. */
-    public val typedSize: StrictlyPositiveIntJvm
-        get() = StrictlyPositiveIntJvm(size)
+    public val typedSize: StrictlyPositiveInt
+        get() = StrictlyPositiveInt(size)
 
     override fun isEmpty(): Boolean = false
 
@@ -80,14 +80,14 @@ public sealed interface NotEmptyCollection<out E> : Collection<E> {
      * throws an [IndexOutOfBoundsException] if the [index] is out of bounds.
      */
     @Throws(IndexOutOfBoundsException::class)
-    public infix operator fun get(index: PositiveIntJvm): E = get(index.value)
+    public infix operator fun get(index: PositiveInt): E = get(index.value)
 
     /**
      * Returns the element at the specified [index] in this collection, or
      * throws an [IndexOutOfBoundsException] if the [index] is out of bounds.
      */
     @Throws(IndexOutOfBoundsException::class)
-    public infix operator fun get(index: StrictlyPositiveIntJvm): E =
+    public infix operator fun get(index: StrictlyPositiveInt): E =
         get(index.value)
 
     /**
@@ -104,14 +104,14 @@ public sealed interface NotEmptyCollection<out E> : Collection<E> {
      * Returns the element at the specified [index] in this collection, or
      * returns `null` if the [index] is out of bounds.
      */
-    public infix fun getOrNull(index: PositiveIntJvm): E? =
+    public infix fun getOrNull(index: PositiveInt): E? =
         getOrNull(index.value)
 
     /**
      * Returns the element at the specified [index] in this collection, or
      * returns `null` if the [index] is out of bounds.
      */
-    public infix fun getOrNull(index: StrictlyPositiveIntJvm): E? =
+    public infix fun getOrNull(index: StrictlyPositiveInt): E? =
         getOrNull(index.value)
 
     // ---------- Conversions ----------